Split tests by drawing type
